+++ app/src/Shared/Shared.fs
@@ -2,6 +2,12 @@ namespace Shared
 
 open System
 
+[<RequireQualifiedAccess>]
+type TodoUpdate =
+    | Create of description: string
+    | Update of id: Guid * description: string
+    | Delete of id: Guid
+
 type Todo = { Id: Guid; Description: string }
 
 module Todo =
@@ -16,6 +22,8 @@ module Todo =
 type ITodosApi = {
     getTodos: unit -> Async<Todo list>
     addTodo: Todo -> Async<Todo list>
-    updateTodo: Todo -> Async<Todo list>
+    updateTodo: TodoUpdate -> Async<Todo list>
     deleteTodo: Guid -> Async<Todo list>
+    deleteTodos: Guid [] -> Async<Todo list>
+    clearTodos: unit -> Async<Result<unit, string>>
 }