+++ src/Fable.Remoting.OpenAPI.Suave/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to Semantic Versioning.
 
 ## [Unreleased]
 
+## 0.0.3 - 2026-03-17
+
+- Updated Core version to 0.0.3 for discriminate union support and OpenAPI example serialization compatibility with Fable.Remoting.Json.
+
 ## 0.0.2 - 2026-03-16
 
 ### Changed